+++ src/test/kotlin/test/de/itemis/mps/gradle/ModelCheckTest.kt
@@ -28,12 +28,15 @@ class ModelCheckTest {
             "plugin-classpath.txt"
         )!!.readText().lines().map { File(it) }
         mpsTestPrjLocation = testProjectDir.newFolder("mps-prj")
-        ProjectHelper().extractTestProject(mpsTestPrjLocation)
         junitFile = File(mpsTestPrjLocation, "junit.xml")
     }
 
+    private fun extractProject(name: String) = ProjectHelper().extractTestProject(name, mpsTestPrjLocation)
+
     @Test
     fun `check model works with latest MPS`() {
+        extractProject("test-project")
+
         settingsFile.writeText(
             """
             rootProject.name = "hello-world"
@@ -82,8 +85,118 @@ class ModelCheckTest {
         Assert.assertEquals(TaskOutcome.SUCCESS, result.task(":checkmodels")?.outcome)
         Assert.assertTrue(junitFile.exists())
     }
+
+    @Test
+    fun `check model fails if errors are found`() {
+        extractProject("test-project-with-errors")
+
+        settingsFile.writeText(
+            """
+            rootProject.name = "hello-world"
+        """.trimIndent()
+        )
+
+        buildFile.writeText(
+            """
+            import java.net.URI
+            buildscript {
+                dependencies {
+                    "classpath"(files(${cp.map { """"${it.invariantSeparatorsPath}"""" }.joinToString()}))
+                }
+            }
+            
+            plugins {
+                id("modelcheck")
+            }
+            
+            repositories {
+                mavenCentral()
+                maven {
+                    url = URI("https://projects.itemis.de/nexus/content/repositories/mbeddr")
+                }
+            }
+            
+            val mps = configurations.create("mps")
+            
+            dependencies {
+                mps("com.jetbrains:mps:2021.1.3")
+            }
+            
+            modelcheck {
+                projectLocation = file("${mpsTestPrjLocation.toPath()}")
+                mpsConfig = mps
+                junitFile = file("${junitFile.absolutePath}")
+            }
+        """.trimIndent()
+        )
+
+        val result = GradleRunner.create()
+            .withProjectDir(testProjectDir.root)
+            .withArguments("checkmodels")
+            .withPluginClasspath(cp)
+            .buildAndFail()
+        Assert.assertEquals(TaskOutcome.FAILED, result.task(":checkmodels")?.outcome)
+        Assert.assertTrue(junitFile.exists())
+    }
+
+    @Test
+    fun `check model works with latest MPS and excluded models`() {
+        extractProject("test-project-with-errors")
+
+        settingsFile.writeText(
+            """
+            rootProject.name = "hello-world"
+        """.trimIndent()
+        )
+
+        buildFile.writeText(
+            """
+            import java.net.URI
+            buildscript {
+                dependencies {
+                    "classpath"(files(${cp.map { """"${it.invariantSeparatorsPath}"""" }.joinToString()}))
+                }
+            }
+            
+            plugins {
+                id("modelcheck")
+            }
+            
+            repositories {
+                mavenCentral()
+                maven {
+                    url = URI("https://projects.itemis.de/nexus/content/repositories/mbeddr")
+                }
+            }
+            
+            val mps = configurations.create("mps")
+            
+            dependencies {
+                mps("com.jetbrains:mps:2021.1.3")
+            }
+            
+            modelcheck {
+                projectLocation = file("${mpsTestPrjLocation.toPath()}")
+                mpsConfig = mps
+                junitFile = file("${junitFile.absolutePath}")
+                excludeModels = listOf("my.solution.with.errors.java")
+            }
+        """.trimIndent()
+        )
+
+        val result = GradleRunner.create()
+            .withProjectDir(testProjectDir.root)
+            .withArguments("checkmodels")
+            .withPluginClasspath(cp)
+            .build()
+        Assert.assertEquals(TaskOutcome.SUCCESS, result.task(":checkmodels")?.outcome)
+        Assert.assertTrue(junitFile.exists())
+    }
+
     @Test
     fun `check model fails with unsupported MPS`() {
+        extractProject("test-project")
+
         settingsFile.writeText(
             """
             rootProject.name = "hello-world"
@@ -132,8 +245,11 @@ class ModelCheckTest {
             .buildAndFail()
         Assert.assertFalse(junitFile.exists())
     }
+
     @Test
     fun `check model works with set MPS version and path`() {
+        extractProject("test-project")
+
         settingsFile.writeText(
             """
             rootProject.name = "hello-world"
@@ -184,6 +300,8 @@ class ModelCheckTest {
 
     @Test
     fun `check model fails with set MPS invalid version and path`() {
+        extractProject("test-project")
+
         settingsFile.writeText(
             """
             rootProject.name = "hello-world"
@@ -231,8 +349,11 @@ class ModelCheckTest {
             .withPluginClasspath(cp)
             .buildAndFail()
     }
+
     @Test
     fun `check model fails with only MPS version set`() {
+        extractProject("test-project")
+
         settingsFile.writeText(
             """
             rootProject.name = "hello-world"